+++ update-logstash.sh
@@ -13,6 +13,13 @@ if [[ ${L_VERSION} = '' ]]; then
   L_VERSION=5.1.2
 fi
 
+# Check if already up to date
+L_CVERSION=`dpkg-query -W -f='${Version}\n' logstash | cut -d : -f2 | cut -d - -f1`
+if [[ "${L_VERSION}" = "${L_CVERSION}" ]]; then
+  echo "Logstash is up to date to ${L_CVERSION} version"
+  exit 0
+fi
+
 # Stop Logstash Daemon
 sudo /bin/systemctl stop logstash.service
